+++ packages/shared/forks/ReactFeatureFlags.www.js
@@ -31,7 +31,6 @@ export const {
   disableSchedulerTimeoutInWorkLoop,
   enableLazyContextPropagation,
   enableSyncDefaultUpdates,
-  enableCapturePhaseSelectiveHydrationWithoutDiscreteEventReplay,
   enableClientRenderFallbackOnHydrationMismatch,
   enableClientRenderFallbackOnTextMismatch,
 } = dynamicFeatureFlags;
@@ -72,6 +71,7 @@ export const enableSelectiveHydration = true;
 
 export const enableLazyElements = true;
 export const enableCache = true;
+export const enableCacheElement = true;
 
 export const disableJavaScriptURLs = true;